requirements.txt
numpy
h5py
addict
ruamel.yaml
cv2
pytorch